eLTER getDataset function

This function allows to obtain the info of dataset provided in DEIMS-SDR.

getDataset(datasetid)

Arguments

datasetid
A character. It is the DEIMS iD of dataset make from DEIMS-SDR website. More information about DEIMS iD in this page.

Value

The output of the function is a `tibble` with main features of the site and the related resources collected by site.

Author

Alessandro Oggioni, phD (2020) oggioni.a@irea.cnr.it

Examples

tDataset <- getDataset(datasetid = "https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef")

#>  Found 1 records... Imported 1 records. Simplifying...

map <- leaflet::leaflet(tDataset) %>% 
 leaflet::addTiles() %>% 
 leaflet::addPolygons()
print(map)
tDataset

#> Simple feature collection with 1 feature and 33 fields
#> Geometry type: POLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 11.88721 ymin: 43.20518 xmax: 15.86426 ymax: 45.91294
#> Geodetic CRS:  WGS 84
#> # A tibble: 1 × 34
#>   title  abstract   keywords uri   type  dateRange.from dateRange.to relatedSite
#> * <chr>  <chr>      <list>   <chr> <chr> <chr>          <lgl>        <list>     
#> 1 LTER … The prese… <df [5 … http… data… 1965-01-01     NA           <df [1 × 4…
#> # … with 26 more variables: contacts.corresponding <list>,
#> #   contacts.creator <list>, contacts.metadataProvider <lgl>,
#> #   observationParameters <list>, observationSpecies <list>, dataPolicy <list>,
#> #   doi <chr>, onlineLocation <list>, legal.accessUse <list>,
#> #   legal.rights <lgl>, legal.legalAct <lgl>, legal.citation <lgl>,
#> #   method.instrumentation <lgl>, method.qualityAssurance <lgl>,
#> #   method.methodUrl <list>, method.methodDescription <list>, …