eLTER_getILTEREnvCharacts

This function ...

getILTEREnvCharacts()

Value

The output of the function is a dataframe containing all the Environmental Characteristics of the all ILTER's sites.

Author

Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>

Examples

listEnvCharacts <- getILTEREnvCharacts()
knitr::kable(listEnvCharacts[1:5, ])

#> 
#> 
#> |title                                                         |uri                                                    |geoCoord                |country | geoElev.avg| geoElev.min| geoElev.max|geoElev.unit | envCharacteristics.airTemperature.avg| envCharacteristics.airTemperature.min| envCharacteristics.airTemperature.max|envCharacteristics.airTemperature.unit | envCharacteristics.precipitation.annual| envCharacteristics.precipitation.min| envCharacteristics.precipitation.max|envCharacteristics.precipitation.unit |envCharacteristics.biogeographicalRegion |envCharacteristics.biome |envCharacteristics.ecosystemAndLanduse |envCharacteristics.eunisHabitat |envCharacteristics.geoBonBiome |envCharacteristics.geology |envCharacteristics.hydrology |envCharacteristics.soils |envCharacteristics.vegetation |
#> |:-------------------------------------------------------------|:------------------------------------------------------|:-----------------------|:-------|-----------:|-----------:|-----------:|:------------|-------------------------------------:|-------------------------------------:|-------------------------------------:|:--------------------------------------|---------------------------------------:|------------------------------------:|------------------------------------:|:-------------------------------------|:----------------------------------------|:------------------------|:--------------------------------------|:-------------------------------|:------------------------------|:--------------------------|:----------------------------|:------------------------|:-----------------------------|
#> |Kosutnjak and Miljakovac (Košutnjak and Miljakovac) - Serbia  |https://deims.org/3e97f1ff-c3ad-4d19-91f9-5ee3485878cd |POINT (20.4619 44.7547) |Serbia  |          NA|          NA|          NA|msl          |                                    NA|                                    NA|                                    NA|°C                                     |                                      NA|                                   NA|                                   NA|mm                                    |continental                              |NA                       |NA                                     |NA                              |NA                             |NA                         |NA                           |NA                       |NA                            |
#> |Pancevacki rit - Serbia                                       |https://deims.org/da850c83-d611-4860-8b78-b663d070655d |POINT (20.5372 44.8761) |Serbia  |          NA|          NA|          NA|msl          |                                    NA|                                    NA|                                    NA|°C                                     |                                      NA|                                   NA|                                   NA|mm                                    |pannonian                                |NA                       |NA                                     |NA                              |NA                             |NA                         |NA                           |NA                       |NA                            |
#> |Lazarev kanjon Nature Reserve - Serbia                        |https://deims.org/238c891e-6690-4b47-ae04-c380b4ae4dca |POINT (21.8828 44.01)   |Serbia  |          NA|          NA|          NA|msl          |                                    NA|                                    NA|                                    NA|°C                                     |                                      NA|                                   NA|                                   NA|mm                                    |continental                              |NA                       |NA                                     |NA                              |NA                             |NA                         |NA                           |NA                       |NA                            |
#> |Landscape of Outstanding Feature Veliko ratno ostrvo - Serbia |https://deims.org/843c85c3-8562-4a6f-949a-bb41d4e27846 |POINT (20.4367 44.8367) |Serbia  |          NA|          NA|          NA|msl          |                                    NA|                                    NA|                                    NA|°C                                     |                                      NA|                                   NA|                                   NA|mm                                    |pannonian                                |NA                       |NA                                     |NA                              |NA                             |NA                         |NA                           |NA                       |NA                            |
#> |Derdap - Serbia                                               |https://deims.org/b13804d7-255b-47c0-86e2-713e33344492 |POINT (22.1 44.5583)    |Serbia  |          NA|          NA|          NA|msl          |                                    NA|                                    NA|                                    NA|°C                                     |                                      NA|                                   NA|                                   NA|mm                                    |continental                              |NA                       |NA                                     |NA                              |NA                             |NA                         |NA                           |NA                       |NA                            |