Return a spatial point vector object containing name and geo-information, stored in DEIMS-SDR catalogue, of all the eLTER sites (more than 1200 sites) belonging to International LTER network (ILTER).

This function downloads generic information of all sites, or a subset of sites of ILTER, through the DEIMS-SDR API. If no `country_name` or `site_name` are specified, the whole list of sites is returned. If either or both of the filtering strings is specified, then a filtered subset of the sites is acquired.

Usage

get_ilter_generalinfo(country_name = NA, site_name = NA, show_map = FALSE)

Arguments

country_name

A `character`. This character string filters the full set of DEIMS sites by country name. Partial matching is supported.

site_name

A `character`. This character string filters by site name where, again, partial matching is supported

show_map

A `boolean`. If TRUE a Leaflet map of site locations is shown. Default FALSE

Value

An `sf` object of the bounding boxes of sites in the filtered list, containing the name, DEIMS.iD, longitude, latitude, average altitude, and affiliation of the filtered ILTER sites. If no bounding box is available,the centroid is returned.

Author

Alessandro Oggioni, phD (2021) oggioni.a@irea.cnr.it

Micha Silver, phD (2021) silverm@post.bgu.ac.il

Examples

if (FALSE) {
# list of the all sites info with ILTER
listOfAllSites <- get_ilter_generalinfo()
length(listOfAllSites[,1])

# example about country name parameter
sitesAustria <- get_ilter_generalinfo(country_name = "Austri")
# (matches Austria, but not Australia)
length(sitesAustria$title)

# example of single site in a country
eisenwurzen <- get_ilter_generalinfo(
  country_name = "Austri",
  site_name = "Eisen"
)
eisenwurzen[,1:2]
# extract DEIMS.Id
eisenwurzen_deimsid <- eisenwurzen$uri
eisenwurzen_deimsid

# example of single site in a country and return only map
get_ilter_generalinfo(
  country_name = "Italy",
  site_name = "Maggiore",
  show_map = TRUE
)
}