Retrieve a list of sites in an eLTER Network.

This function return a spatial point vector object including title, date late updated, URI, and coordinates, stored in DEIMS-SDR catalogue, of all the eLTER sites belonging to an eLTER Network (e.g. LTER- Italy network).

Usage

get_network_sites(networkDEIMSID)

Arguments

networkDEIMSID

A `character`. The DEIMS ID of the network from DEIMS-SDR website. DEIMS ID information here and Complete list of networks here. The DEIMS ID of network is the URL for the network page.

Value

The output of the function is a point vector of `sf` class (package sf) of the network's sites.

Author

Alessandro Oggioni, phD (2020) oggioni.a@irea.cnr.it

Examples

if (FALSE) {
# The sites of LTER-Italy network
listSites <- get_network_sites(
  networkDEIMSID =
  "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listSites

# The sites of LTER Europe network
euSites <- get_network_sites(
  networkDEIMSID =
  "https://deims.org/networks/4742ffca-65ac-4aae-815f-83738500a1fc"
)
euSites
}