Obtain the observed properties measured by a sensor.

This function obtains the observed properties by procedure/sensor through Sensor Observation Service (SOS).

Usage

get_sensor_observed_properties(sosURL, procedure)

Arguments

sosURL

A character. The endpoint of the Sensor Observation Service (SOS) service.

procedure

A character. It is a procedure/sensor ID.

Value

The output of the function is a tibble with the labels and URI (Uniform Resource Identifier) of each observed property, the code and URI of Units Of Measurement (UOM) of the observed properties as declared in the Sensor Observation Service (SOS). Codes and URIs as stated in QUDT.org are also present. QUDT is a public charity nonprofit organization founded to provide semantic specifications for units of measure, quantity kind, dimensions and data types. NB this function returns a valued string only in the case where the UOM refers to a NERC vocabulary term (e.g. http://vocab.nerc.ac.uk/collection/P06/current/UPAA/ for °C).

Author

Alessandro Oggioni, phD oggioni.a@irea.cnr.it

Examples

if (FALSE) {
FP <- get_sensor_observed_properties(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/1286194C-A5DF-11DF-8ED7-1602DFD72097"
)
FP

eurac_monalisa <- get_sensor_observed_properties(
  sosURL = "http://monalisasos.eurac.edu/sos/service",
  procedure = "QuantumSensor_nemef2000"
)
eurac_monalisa

obsProsAir <- get_sensor_observed_properties(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/SI000049-1675AirTemp"
)
obsProsAir

NIVA <- get_sensor_observed_properties(
  sosURL = "https://hydro-sos.niwa.co.nz/",
  procedure = "Hydrometric_Station"
)
NIVA
}