eLTER get_site_MODIS function

This function acquires MODIS products from MODIS Land Products archive, crops to an eLTER site boundary, and plots a time series of the MODIS derived variable averaged over the eLTER site.

get_site_MODIS(
  deimsid,
  earthdata_user,
  earthdata_passwd,
  product = "Vegetation_Indexes_Monthly_1Km (M*D13A3)",
  from_date = NA,
  to_date = NA,
  bands = c("NDVI", "EVI"),
  scale = TRUE,
  save_ts_dir = NA,
  out_folder = tempdir(),
  show_products = FALSE,
  show_bands = NA
)

Arguments

deimsid

a `character`. The DEIMS ID of the site from DEIMS-SDR website. More information about DEIMS ID from: page.

earthdata_user

`character`. For authentication. Please create an account on: https://urs.earthdata.nasa.gov/users/new and enter authentication details here.

earthdata_passwd

`character` For authentication on Earthdata.

product

a `character`. The requested MODIS product. Must be the full product name. Use `show_products = TRUE` to get a list of products. Some examples: "Vegetation_Indexes_Monthly_1Km (M*D13A3)" "LST_3band_emissivity_8day_1km (M*D21A2)" "LAI_8Days_500m (MCD15A2H)" Default is "Vegetation_Indexes_Monthly_1Km (M*D13A3)"

from_date,

`character`, the start date for acquiring MODIS formatted as YYYY.MM.DD Must be later than 2000.02.01. (The earliest MODIS imagery was Feb 2000)

to_date,

`character`, the last date for acquiring MODIS formatted as YYYY.MM.DD.

bands

`character`, string array of bands in the requested product. Use the "show_bands" parameter first to get a list of available bands. Some examples: "LST_Day_1km", or multiple bands as c("NDVI", "EVI") Default is c("NDVI", "EVI").

scale

`boolean`. The MODIS raster values will be scaled (multiplied) by a factor. Some MODIS products are scaled up and saved as 16 bit integers. For example, Vegetation indices are scaled by 10000, and Land surface temperatures are scaled by 50. Scaling factors are derived from product metadata. For additional details refer to the MODIS Product Tables: https://modis.gsfc.nasa.gov/data/dataprod/ Default TRUE (scaling applied)

save_ts_dir,

`character`. Full path to directory to save a CSV file of the time series of site averaged MODIS bands values. Default is NA (saving is disabled). This directory is created if it does not exist.

out_folder,

`character`, Where to save the downloaded MODIS rasters Default is tempdir(). (Note that if this parameter is left at default the downloaded MODIS rasters will be deleted on exit, along with other temporary R files.)

show_products,

a `boolean`, if TRUE, then show a list of all available MODIS products, and exit. Default is FALSE.

show_bands,

a `character`, the name of a MODIS product. The parameter could be one of: "M*D13Q1", (Vegetation Indexes_16Days_250m), "M*D21A2", (LST_3band_emissivity_8day_1km), "MCD12Q1". (LandCover_Type_Yearly_500m) When this parameter is supplied, show a list of the bands for that product, and exit. Default is NA.

Value

The function returns a SpatRaster object (from the `terra` package) which contains a stack of all rasters for the requested product/bands over the requested date interval. The user should save the raster stack to disk, if necessary. i.e. writeRaster(ds_site, "site_MODIS_dataset.tif") The function also plots a time series graph of the site averaged values.

Details

The Nasa [EarthData](https://urs.earthdata.nasa.gov/home) site offers MODIS products as 8 day, 16 day and monthly aggregations. This function accesses those aggregations selecting a subset from the `from_date` to the `to_date`, then prepares site averaged values of the MODIS product, and plots a time series of those averages. In order to download from EarthData, you must register at https://urs.earthdata.nasa.gov/users/new and supply your username and password to this function.

Author

Micha Silver, phD (2020) silverm@post.bgu.ac.il

Alessandro Oggioni, phD (2020) oggioni.a@irea.cnr.it

Examples

 if (FALSE) {
 deimsid <-  "https://deims.org/bf2cb8dd-0d38-419b-aa9c-7d53337bd98e"
 # Use MODIS Land Surface Temperature product
 product <-  "LST_3band_emissivity_8day_1km (M*D21A2)"
 # Check what bands are available for that product
 get_site_MODIS(show_bands = "LST_3band_emissivity_8day_1km (M*D21A2)")
 # Use LST_Day_1KM and LST_Night_1KM
 bands = c("LST_Day_1KM", "LST_Night_1KM")
 # Acquire MODIS data
 # (IMPORTANT: define personal authentication credentials)
 MODIS_LST <- get_site_MODIS(deimsid = deimsid,
                            earthdata_user = "your_USGS_username",
                           earthdata_passwd = "your_USGS_password",
                           product = product,
                           bands = bands,
                           from_date="2021.06.01", to_date = "2021.07.31")
 terra::nlyr(MODIS_LST)
 # Show time series table
 MODIS_LST_timeseries <- read.csv(
   file.path(tempdir(),
           "Time_Series",
           "MODIS_LST_Day_1KM_2021-06-02_2021-07-28.csv"))
 MODIS_LST_timeseries
}