Compose an eLTER Data Reporting Format object

Given several tables, creates an eLTER data reporting format object

Usage

reporting_produce_data_object_v1.3(
  data = NULL,
  station = NULL,
  method = NULL,
  reference = NULL,
  event = NULL,
  sample = NULL,
  licence = "",
  deimsid = "",
  data_type = "measurement",
  filename = NULL
)

Arguments

data

A tibble. See eLTER data specification format for details

station

A tibble

method

A tibble

reference

A tibble

event

A tibble

sample

A tibble

licence

A character

deimsid

A character. The DEIMS ID of the site from DEIMS-SDR website. DEIMS ID information here.

data_type

A character

filename

optional filename associated with the object, of the form provided as output by the function reporting_compose_file_name

Value

list with eLTER reporting format slots

Note

This method must be intended as a signpost for future implementation

See also

Author

Paolo Tagliolato, phD tagliolato.p@irea.cnr.it

Alessandro Oggioni, phD oggioni.a@irea.cnr.it

Examples

if (FALSE) {
## Not run:

deimsid <- "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6"
time_span <- 2015 # e.g. whole year
# time_span <- "20150302-20180415" # e.g. span between two dates
data_topic <- "VEG" # data provider defined abbreviation of "vegetation"
variable_group <- "SPECCOVER" # data provider defined abbreviation
version <- "V20220907"

filename <- reporting_compose_file_name(
  deimsid = deimsid,
  data_topic = data_topic,
  variable_group = variable_group,
  time_span = time_span,
  version = version
)

data <- dplyr::tribble(
  ~id, ~value,
  1, 7.5,
  2, 4.2
)
station <- dplyr::tribble(
  ~SITE_CODE, ~STATION_CODE, ~STYPE, ~LAT,      ~LON,       ~ALTITUDE,
  deimsid,    "IP2",         "AREA",  45.340805, 7.88887495, 265
)
method <- dplyr::tribble(
  ~VARIABLE, ~METH_DESCR,
  "COVE_F",  "Analysis of ammonium..." 
)

research_object <- reporting_produce_data_object_v1.3(
 filename = filename,
 deimsid = deimsid,
 data = data,
 station = station,
 method = method
)

}
## End (Not run)