eLTER taxon_id_worms function

This function enriches and certifies a list of species names by comparing with Worms records

taxon_id_worms(input, taxaColumn = 1, verbose = TRUE, refine = FALSE)

Arguments

taxaColumn
a `numeric`. The cardinal number of the column where species list is. Default is `1`.

verbose
a `logical`. Whit this selection, the function returns a message with number of record(s) that don't match with any Worms names and the number of record(s) that match with more that one Worms name. Default is `TRUE`.

refine
a `logical`. With this selection, the function allows to refine the result(s) that match with more Worms records. By a interactive use of the terminal, the user can chose the result. Default is `FALSE`.

table
a `tibble`. The table that contain the species names list to be checked.

Value

The output of the function is a `tibble` with the columns provided and new columns such as: valid_name, valid_authority, valid_AphiaID, status, synonyms, LSID, url, matchType, nOfWormsRecords, wormsRecords obtained by Worms rest API. The function also return, if verbose is TRUE, the list of records that don't match with Worms name species.

Author

Alessandro Oggioni, phD (2021) oggioni.a@irea.cnr.it

Paolo Tagliolato, phD (2021) tagliolato.p@irea.cnr.it

Examples

phytoplankton <- tibble::tibble(
   ID = c(1, 2, 3, 4, 5, 6, 7),
   species = c("Asterionella formosa", "Chrysococcus sp.", "Cryptomonas rostrata", "Dinobryon divergens", "Mallomonas akrokomos", "Melosira varians", "Cryptomonas rostrata")
)
table <- taxon_id_worms(input = phytoplankton, taxaColumn = 2, verbose = TRUE, refine = TRUE)
#> 
#> After the finding operation the number of record(s) that don't match with any Worms names are:
#> *--- 1 on 7 examined ---*
#> please verify the species name provided and run again this function. The record(s) that match with more that one Worms name are:
#> *--- 2 on 7 examined ---*
#> please use the function taxon_id_worms_refine for specify wich is the exact corrispondence with your given species name.
#> This is the taxa name provided by you:
#> Cryptomonas rostrata
#> Worms don't contain a unique records that match with this name.
#> The Worms records most similar are:
#> 
#> 1:  Cryptomonas rostrata (Skuja, 1948)
#>     Worms status: accepted
#>     Unaccept reason: NA
#>     Match type: exact
#>     Modified: 2015-06-26T12:14:04.327Z
#> 2:  Cryptomonas rostrata (O.V.Troitzkaja, 1922)
#>     Worms status: unaccepted
#>     Unaccept reason: synonym
#>     Match type: exact
#>     Modified: 2021-08-27T06:27:33.180Z
#> 
#> Please select the record that you think most similar to the taxa name that you have provided. Insert the number of record: 
#> This is the taxa name provided by you:
#> Cryptomonas rostrata
#> Worms don't contain a unique records that match with this name.
#> The Worms records most similar are:
#> 
#> 1:  Cryptomonas rostrata (Skuja, 1948)
#>     Worms status: accepted
#>     Unaccept reason: NA
#>     Match type: exact
#>     Modified: 2015-06-26T12:14:04.327Z
#> 2:  Cryptomonas rostrata (O.V.Troitzkaja, 1922)
#>     Worms status: unaccepted
#>     Unaccept reason: synonym
#>     Match type: exact
#>     Modified: 2021-08-27T06:27:33.180Z
#> 
#> Please select the record that you think most similar to the taxa name that you have provided. Insert the number of record: 
table
#> # A tibble: 7 × 12
#>      ID species  valid_name valid_authority valid_AphiaID status synonyms LSID  
#>   <dbl> <chr>    <chr>      <chr>                   <int> <chr>  <chr>    <chr> 
#> 1     1 Asterio… Asterione… Hassall, 1850          148954 accep… NA       urn:l…
#> 2     2 Chrysoc… NA         NA                         NA NA     NA       NA    
#> 3     3 Cryptom… NA         NA                         NA NA     NA       NA    
#> 4     4 Dinobry… Dinobryon… O.E.Imhof, 1887        157248 accep… NA       urn:l…
#> 5     5 Mallomo… Mallomona… Ruttner, 1913          249722 accep… NA       urn:l…
#> 6     6 Melosir… Melosira … C.Agardh, 1827         149043 accep… NA       urn:l…
#> 7     7 Cryptom… NA         NA                         NA NA     NA       NA    
#> # … with 4 more variables: url <chr>, matchType <chr>, nOfWormsResults <dbl>,
#> #   wormsRecords <list>