Enrich and certify a list of species names by comparing with GBIF.

This function provide tibble object with all the columns of input table of taxa plus new columns obtained from: GBIF Backbone Taxonomy. This function use the rgbif::name_backbone_checklist function for interact with API backbone taxonomy of GBIF.

Usage

taxon_id_gbif(input, taxaColumn = 1, refine = FALSE)

Arguments

input

A tibble. The table that contain the species names list to be checked.

taxaColumn

A numeric. The cardinal number of the column where species list is. Default is 1.

refine

A logical. With this selection, the function allows to refine the result(s) that match with more GBIF records. By a interactive use of the terminal, the user can chose the result. Default is FALSE.

Value

When the refine is FALSE the output of the function is a tibble with the refinement of the input, else the function simply returns the unchanged input. All the labels of the columns are the terms of Darwin Core terms. The columns labels are annotate with the link (URI) of the Darwin Core termsas attributes of the tibble.

References

Chamberlain S, Barve V, Mcglinn D, Oldoni D, Desmet P, Geffert L, Ram K (2023). rgbif: Interface to the Global Biodiversity Information Facility API. R package version 3.7.5, https://CRAN.R-project.org/package=rgbif.

Wickham H, François R, Henry L, Müller K (2022). dplyr: A Grammar of Data Manipulation. R package version 1.0.9, https://CRAN.R-project.org/package=dplyr.

Müller K, Wickham H (2022). tibble: Simple Data Frames. R package version 3.1.7, https://CRAN.R-project.org/package=tibble.

Author

Alessandro Oggioni, phD (2021) oggioni.a@irea.cnr.it

Examples

macrophytes <- tibble::tibble(
  ID = c(1:18),
  species = c("Ceratophyllum demersum",
  "Lagarosiphon major",
  "Littorella uniflora",
  "Myriophyllum spicatum",
  "Myriophyllum verticillatum",
  "Najas marina",
  "Najas minor",
  "Nelumbo nucifera",
  "Nitella gracilis",
  "Nuphar lutea",
  "Nymphaea alba",
  "Nymphoides peltata",
  "Persicaria amphibia",
  "Potamogeton perfoliatus",
  "Potamogeton pusillus",
  "Ranunculus sp.",
  "Stuckenia pectinata",
  "Trapa natans")
)
table <- taxon_id_gbif(
  input = macrophytes,
  taxaColumn = 2,
  refine = TRUE
)
#> Error in taxon_id_gbif(input = macrophytes, taxaColumn = 2, refine = TRUE): could not find function "taxon_id_gbif"
table
#> function (..., exclude = if (useNA == "no") c(NA, NaN), useNA = c("no", 
#>     "ifany", "always"), dnn = list.names(...), deparse.level = 1) 
#> {
#>     list.names <- function(...) {
#>         l <- as.list(substitute(list(...)))[-1L]
#>         nm <- names(l)
#>         fixup <- if (is.null(nm)) 
#>             seq_along(l)
#>         else nm == ""
#>         dep <- vapply(l[fixup], function(x) switch(deparse.level + 
#>             1, "", if (is.symbol(x)) as.character(x) else "", 
#>             deparse(x, nlines = 1)[1L]), "")
#>         if (is.null(nm)) 
#>             dep
#>         else {
#>             nm[fixup] <- dep
#>             nm
#>         }
#>     }
#>     miss.use <- missing(useNA)
#>     miss.exc <- missing(exclude)
#>     useNA <- if (miss.use && !miss.exc && !match(NA, exclude, 
#>         nomatch = 0L)) 
#>         "ifany"
#>     else match.arg(useNA)
#>     doNA <- useNA != "no"
#>     if (!miss.use && !miss.exc && doNA && match(NA, exclude, 
#>         nomatch = 0L)) 
#>         warning("'exclude' containing NA and 'useNA' != \"no\"' are a bit contradicting")
#>     args <- list(...)
#>     if (length(args) == 1L && is.list(args[[1L]])) {
#>         args <- args[[1L]]
#>         if (length(dnn) != length(args)) 
#>             dnn <- if (!is.null(argn <- names(args))) 
#>                 argn
#>             else paste(dnn[1L], seq_along(args), sep = ".")
#>     }
#>     if (!length(args)) 
#>         stop("nothing to tabulate")
#>     bin <- 0L
#>     lens <- NULL
#>     dims <- integer()
#>     pd <- 1L
#>     dn <- NULL
#>     for (a in args) {
#>         if (is.null(lens)) 
#>             lens <- length(a)
#>         else if (length(a) != lens) 
#>             stop("all arguments must have the same length")
#>         fact.a <- is.factor(a)
#>         if (doNA) 
#>             aNA <- anyNA(a)
#>         if (!fact.a) {
#>             a0 <- a
#>             a <- factor(a, exclude = exclude)
#>         }
#>         add.na <- doNA
#>         if (add.na) {
#>             ifany <- (useNA == "ifany")
#>             anNAc <- anyNA(a)
#>             add.na <- if (!ifany || anNAc) {
#>                 ll <- levels(a)
#>                 if (add.ll <- !anyNA(ll)) {
#>                   ll <- c(ll, NA)
#>                   TRUE
#>                 }
#>                 else if (!ifany && !anNAc) 
#>                   FALSE
#>                 else TRUE
#>             }
#>             else FALSE
#>         }
#>         if (add.na) 
#>             a <- factor(a, levels = ll, exclude = NULL)
#>         else ll <- levels(a)
#>         a <- as.integer(a)
#>         if (fact.a && !miss.exc) {
#>             ll <- ll[keep <- which(match(ll, exclude, nomatch = 0L) == 
#>                 0L)]
#>             a <- match(a, keep)
#>         }
#>         else if (!fact.a && add.na) {
#>             if (ifany && !aNA && add.ll) {
#>                 ll <- ll[!is.na(ll)]
#>                 is.na(a) <- match(a0, c(exclude, NA), nomatch = 0L) > 
#>                   0L
#>             }
#>             else {
#>                 is.na(a) <- match(a0, exclude, nomatch = 0L) > 
#>                   0L
#>             }
#>         }
#>         nl <- length(ll)
#>         dims <- c(dims, nl)
#>         if (prod(dims) > .Machine$integer.max) 
#>             stop("attempt to make a table with >= 2^31 elements")
#>         dn <- c(dn, list(ll))
#>         bin <- bin + pd * (a - 1L)
#>         pd <- pd * nl
#>     }
#>     names(dn) <- dnn
#>     bin <- bin[!is.na(bin)]
#>     if (length(bin)) 
#>         bin <- bin + 1L
#>     y <- array(tabulate(bin, pd), dims, dimnames = dn)
#>     class(y) <- "table"
#>     y
#> }
#> <bytecode: 0x7fce34948f28>
#> <environment: namespace:base>

# The annotated URIs of columns label are achieved by:
attributes(table)$uri
#> NULL