Enrich and certify a list of species names by comparing with Worms.

This function tibble object with all the columns of input table of taxa plus new columns such as valid_name, valid_authority, valid_AphiaID, status, synonyms, LSID, url, matchType, nOfWormsRecords, wormsRecords obtained from: Worms rest API.

Usage

taxon_id_worms(input, taxaColumn = 1, verbose = TRUE, refine = FALSE)

Arguments

input

A `tibble`. The table that contain the species names list to be checked.

taxaColumn

A `numeric`. The cardinal number of the column where species list is. Default is `1`.

verbose

A `logical`. Whit this selection, the function returns a message with number of record(s) that don't match with any Worms names and the number of record(s) that match with more that one Worms name. Default is `TRUE`.

refine

A `logical`. With this selection, the function allows to refine the result(s) that match with more Worms records. By a interactive use of the terminal, the user can chose the result. Default is `FALSE`.

Value

The output of the function is a `tibble` with the columns provided and new columns such as: valid_name, valid_authority, valid_AphiaID, status, synonyms, LSID, url, matchType, nOfWormsRecords, wormsRecords obtained by Worms rest API. The function also return, if verbose is TRUE, the list of records that don't match with Worms name species.

Author

Alessandro Oggioni, phD (2021) oggioni.a@irea.cnr.it

Paolo Tagliolato, phD (2021) tagliolato.p@irea.cnr.it

Examples

phytoplankton <- tibble::tibble(
   ID = c(1, 2, 3, 4, 5, 6, 7),
   species = c(
   "Asterionella formosa", "Chrysococcus sp.",
   "Cryptomonas rostrata", "Dinobryon divergens",
   "Mallomonas akrokomos", "Melosira varians",
   "Cryptomonas rostrata"
 )
)
table <- taxon_id_worms(
  input = phytoplankton,
  taxaColumn = 2,
  verbose = TRUE,
  refine = TRUE
)
table